Arreglados errores menores. Agregado una imagen de carga mientras trae los personajes por default.
